Implement separate routing for each tab
- Create individual pages for /generate, /prompts, and /analytics routes
- Add shared Layout component with header and footer
- Update navigation to use Next.js Link components for proper routing
- Redirect root path to /generate as default
- Update root layout to use shared Layout component
- Each tab now has its own URL and can be bookmarked/shared
- Maintain state management within individual pages
