+++ apps/frontend/src/app/analytics/page.tsx
@@ -1,0 +1,11 @@
+import AnalyticsDashboard from '@/components/AnalyticsDashboard'
+
+export default function AnalyticsPage() {
+  return (
+    <div className="py-8 px-4 sm:px-6 lg:px-8">
+      <div className="max-w-7xl mx-auto">
+        <AnalyticsDashboard />
+      </div>
+    </div>
+  )
+}
